refactor: removed private methods and created parser + normalizer classes
